Added conditional rendering and additional modularity
