+++ client/src/App.js
@@ -1,20 +1,15 @@
 import React, { useState, useEffect } from 'react';
 import { BrowserRouter, Route } from 'react-router-dom';
-import { Provider } from 'react-redux';
+import { useDispatch } from 'react-redux';
+import { setUser } from './store/auth';
 import HomePage from './components/HomePage';
 import PageLoad from './components/PageLoad';
 import LogIn from './components/LogIn';
 import SignUp from './components/SignUp';
-import configureStore from './store/configureStore';
-
-export const store = configureStore();
-
-if (process.env.NODE_ENV !== 'production') {
-  window.store = store;
-}
 
 function App() {
   const [loading, setLoading] = useState(true);
+  const dispatch = useDispatch();
 
   useEffect(() => {
     const loadUser = async () => {
@@ -22,13 +17,14 @@ function App() {
       const res = await fetch("/api/session");
       if (res.ok) {
         res.data = await res.json(); // current user info
+        dispatch(setUser(res.data.user));
       }
       setTimeout(() => {
         setLoading(false);
       }, 1200)
     }
     loadUser();
-  }, []);
+  }, [dispatch]);
 
   if (loading) return (
     <PageLoad />
@@ -36,17 +32,15 @@ function App() {
 
     return (
       <BrowserRouter>
-      <Provider store={store}>
         <Route exact path='/'>
           <HomePage />
         </Route>
         <Route path="/login">
-          <LogIn store={store}/>
+          <LogIn />
         </Route>
         <Route path='/sign-up'>
-          <SignUp store={store}/>
+          <SignUp />
         </Route>
-      </Provider>
       </BrowserRouter>
     );
 }
