+++ client/src/components/HomePage.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import { useSelector } from 'react-redux';
-import { MDBContainer, MDBBox, MDBDropdown, MDBDropdownMenu, MDBDropdownToggle } from 'mdbreact';
+import { MDBContainer, MDBBox } from 'mdbreact';
 import styles from '../css-modules/HomePage.module.css';
-import LogIn from './LogIn';
-import SignUp from './SignUp';
+import LogInDropDown from './LogInDropDown';
+import SignUpDropDown from './SignUpDropDown';
 
 const HomePage = () => {
     const currentUser = useSelector(state => state.auth.id);
@@ -16,24 +16,14 @@ const HomePage = () => {
                         <img className={styles.flowNotesLogo} src='/images/flownotesLogo.png' alt='Marvel vs. Capcom cast'/>
                         <span className={styles.flowNotesText}>flowNotes</span>
                     </a>
-                    <div className={styles.headerTopRow__rightContainer}>
-                        <MDBDropdown>
-                            <MDBDropdownToggle caret className={styles.dropdownButton} color='amber darken-4'>
-                                Login
-                            </MDBDropdownToggle>
-                            <MDBDropdownMenu>
-                                <LogIn />
-                            </MDBDropdownMenu>
-                        </MDBDropdown>
-                        <MDBDropdown>
-                            <MDBDropdownToggle caret className={styles.dropdownButton} color='amber darken-4'>
-                                Sign Up
-                            </MDBDropdownToggle>
-                            <MDBDropdownMenu>
-                                <SignUp />
-                            </MDBDropdownMenu>
-                        </MDBDropdown>
-                    </div>
+                        {
+                            currentUser ?
+                            <div></div> :
+                            <div className={styles.headerTopRow__rightContainer_notLoggedIn}>
+                            <LogInDropDown />
+                            <SignUpDropDown />
+                            </div>
+                        }
                 </div>
             </MDBBox>
             </div>
