+++ client/src/css-modules/LoginForm.module.css
@@ -1,7 +1,7 @@
 .submitButton {
-    margin-left: 3.7em;
+    margin-left: 5.5em;
 }
 
 .loginForm {
-    width: 10vw;
+    width: 12vw;
 }
